rename components: PluginRegistryPane -> PluginRegistryTab; PluginRegistryDetailPane -> PluginDetailPane
